refactor: use m'/44'/626'/<index>
