+++ packages/libs/hd-wallet/src/SLIP10/index.ts
