+++ packages/libs/hd-wallet/src/SLIP10/kadenaGenKeypairFromSeed.ts
@@ -39,7 +39,7 @@ export function kadenaGenKeypairFromSeed(
 
 /**
  * Generates a key pair from a seed buffer and an index or range of indices, and optionally encrypts the private key.
- * it uses bip44 m'/44'/626'/${index}'/0'/0' derivation path
+ * it uses bip44 m'/44'/626'/${index}' derivation path
  *
  * @param {Uint8Array} seedBuffer - The seed buffer to use for key generation.
  * @param {number | [number, number]} indexOrRange - Either a single index or a tuple with start and end indices for key pair generation.
@@ -51,7 +51,7 @@ export function kadenaGenKeypairFromSeed(
   password: string,
   seed: EncryptedString,
   indexOrRange: number | [number, number],
-  derivationPathTemplate: string = `m'/44'/626'/<index>'/0'/0'`,
+  derivationPathTemplate: string = `m'/44'/626'/<index>'`,
 ): [string, EncryptedString] | Array<[string, EncryptedString]> {
   if (typeof seed !== 'string' || seed === '') {
     throw new Error('No seed provided.');
