+++ packages/libs/hd-wallet/src/SLIP10/kadenaMnemonic.ts
