+++ packages/libs/hd-wallet/src/SLIP10/tests/kadenaGetPublic.test.ts
@@ -41,15 +41,20 @@ describe('kadenaGetPublic', () => {
       // this mnemonic is generated by Enkrypt wallet
       'coyote utility final warfare thumb symbol mule scale final nominee behave crumble',
     );
-    let publicKey = kadenaGetPublic(password, seed, 0);
+    let publicKey = kadenaGetPublic(
+      password,
+      seed,
+      0,
+      "m'/44'/626'/<index>/0'/0'",
+    );
     expect(publicKey).toBe(
       '43726c4a2e7b03fa5d23635307e5b130baf8b261e1081c099a2b43db1d4554cc',
     );
-    publicKey = kadenaGetPublic(password, seed, 1);
+    publicKey = kadenaGetPublic(password, seed, 1, "m'/44'/626'/<index>/0'/0'");
     expect(publicKey).toBe(
       '3f53dfad097fdf8501c32b275e109980ed7121866a63ca34bb035c4a2e41a265',
     );
-    publicKey = kadenaGetPublic(password, seed, 2);
+    publicKey = kadenaGetPublic(password, seed, 2, "m'/44'/626'/<index>/0'/0'");
     expect(publicKey).toBe(
       '3021bcfa703cc4fac007ab4c5050df5c0b8ca7d655ea80c84af9ea5e43ecf0ff',
     );
