+++ packages/libs/hd-wallet/src/SLIP10/tests/kadenaSignWithKeyPair.test.ts
