+++ packages/libs/hd-wallet/src/SLIP10/tests/kadenaSignWithSeed.test.ts
