+++ packages/libs/hd-wallet/src/SLIP10/utils/sign.ts
